refactor(ForumTopicController): migrate recentPosts logic to an action
