+++ app/Community/Controllers/ForumTopicController.php
@@ -4,24 +4,23 @@ declare(strict_types=1);
 
 namespace App\Community\Controllers;
 
+use App\Community\Actions\BuildAggregateRecentForumPostsDataAction;
 use App\Community\Data\RecentPostsPagePropsData;
 use App\Community\Requests\ForumTopicRequest;
-use App\Data\ForumTopicData;
-use App\Data\PaginatedData;
 use App\Enums\Permissions;
+use App\Http\Controller;
 use App\Models\Forum;
 use App\Models\ForumTopic;
 use App\Models\User;
 use Illuminate\Contracts\View\View;
 use Illuminate\Http\RedirectResponse;
 use Illuminate\Http\Request;
-use Illuminate\Pagination\LengthAwarePaginator;
 use Illuminate\Support\Collection;
 use Illuminate\Support\Facades\Auth;
 use Inertia\Inertia;
 use Inertia\Response as InertiaResponse;
 
-class ForumTopicController extends \App\Http\Controller
+class ForumTopicController extends Controller
 {
     public function index(): void
     {
@@ -101,123 +100,24 @@ class ForumTopicController extends \App\Http\Controller
         $this->authorize('delete', $topic);
     }
 
-    private function getTotalRecentForumTopics(int $permissions = Permissions::Unregistered): int
-    {
-        return ForumTopic::query()
-            ->where("RequiredPermissions", "<=", $permissions)
-            ->whereNull("deleted_at")
-            ->where(function ($query) {
-                $query
-                    ->whereNotNull("LatestCommentID")
-                    ->orWhereIn("ID", function ($subQuery) {
-                        $subQuery
-                            ->select("ForumTopicID")
-                            ->distinct()
-                            ->from("ForumTopicComment")
-                            ->where("Authorised", 1);
-                        }
-                    );
-                }
-            )
-            ->count();
-    }
-
-    public function recentPosts(Request $request): InertiaResponse
-    {
-        $offset = $request->input('page.number', 1) - 1;
-        $count = 25;
-
-        /** @var User $user */
+    public function recentPosts(
+        Request $request,
+        BuildAggregateRecentForumPostsDataAction $buildAggregateRecentPostsData
+    ): InertiaResponse {
+        /** @var ?User $user */
         $user = Auth::user();
-        $permissions = Permissions::Unregistered;
-        if ($user) {
-            $permissions = (int) $user->getAttribute('Permissions');
-        }
+        $permissions = $user ? (int) $user->getAttribute('Permissions') : Permissions::Unregistered;
 
-        $topics = $this->getRecentForumTopics(
-            page: (int) $request->input('page.number', 1),
+        $paginatedTopics = $buildAggregateRecentPostsData->execute(
             permissions: $permissions,
+            page: (int) $request->input('page.number', 1),
+            limit: 25,
+            paginationPath: $request->url(),
+            paginationQuery: $request->query(),
         );
-
-        $transformedTopics = array_map(
-            fn ($topic) => ForumTopicData::fromRecentlyActiveTopic($topic)->include(
-                'commentCount24h',
-                'oldestComment24hId',
-                'commentCount7d',
-                'oldestComment7dId',
-            ),
-            $topics
-        );
-
-        $paginator = new LengthAwarePaginator(
-            items: $transformedTopics,
-            total: $this->getTotalRecentForumTopics($permissions),
-            perPage: $count,
-            currentPage: $offset + 1,
-            options: [
-                'path' => $request->url(),
-                'query' => $request->query(),
-            ],
-        );
-
-        $paginatedTopics = PaginatedData::fromLengthAwarePaginator($paginator);
 
         $props = new RecentPostsPagePropsData($paginatedTopics);
 
         return Inertia::render('forums/recent-posts', $props);
-    }
-
-    private function getRecentForumTopics(int $page = 1, int $permissions = Permissions::Unregistered): array
-    {
-        $count = 25;
-        $offset = ($page - 1) * $count;
-
-        $query = "
-            SELECT 
-                ft.ID AS ForumTopicID, 
-                ft.Title AS ForumTopicTitle,
-                f.ID AS ForumID, 
-                f.Title AS ForumTitle,
-                lftc.ID AS CommentID, 
-                lftc.DateCreated AS PostedAt, 
-                lftc.author_id,
-                ua.User AS Author, 
-                ua.display_name AS AuthorDisplayName,
-                LEFT(lftc.Payload, 260) AS ShortMsg,
-                LENGTH(lftc.Payload) > 260 AS IsTruncated,
-                d1.CommentID AS CommentID_1d,
-                d1.Count AS Count_1d,
-                d7.CommentID AS CommentID_7d,
-                d7.Count AS Count_7d
-            FROM (
-                SELECT ft.ID, ft.Title, ft.ForumID, ft.LatestCommentID
-                FROM ForumTopic ft
-                FORCE INDEX (idx_permissions_deleted_latest)
-                WHERE ft.RequiredPermissions <= :permissions AND ft.deleted_at IS NULL
-                ORDER BY ft.LatestCommentID DESC
-            ) AS ft
-            INNER JOIN Forum AS f ON f.ID = ft.ForumID
-            INNER JOIN ForumTopicComment AS lftc ON lftc.ID = ft.LatestCommentID AND lftc.Authorised = 1
-            LEFT JOIN UserAccounts AS ua ON ua.ID = lftc.author_id
-            LEFT JOIN (
-                SELECT ForumTopicId, MIN(ID) AS CommentID, COUNT(*) AS Count
-                FROM ForumTopicComment
-                WHERE Authorised = 1 AND DateCreated >= NOW() - INTERVAL 1 DAY
-                GROUP BY ForumTopicId
-            ) AS d1 ON d1.ForumTopicId = ft.ID
-            LEFT JOIN (
-                SELECT ForumTopicId, MIN(ID) AS CommentID, COUNT(*) AS Count
-                FROM ForumTopicComment
-                WHERE Authorised = 1 AND DateCreated >= NOW() - INTERVAL 7 DAY
-                GROUP BY ForumTopicId
-            ) AS d7 ON d7.ForumTopicId = ft.ID
-            ORDER BY lftc.DateCreated DESC
-            LIMIT :offset, :count";
-
-        return legacyDbFetchAll($query, [
-            'offset' => $offset,
-            'count' => $count,
-            'permissions' => $permissions,
-        ])->toArray();
     }
 }
